Adding new labels to pve_up and general metrics.
The labels are all the same for every type of resource, node, cluster,
storage, vm, container; where a key is not applicable, the value is
empty. Other keys like id and name are sythesized where not applicable.
The labels that are always present are:
- id (previously, `id` was the only label always present)
- name
- up
- node
- template
- type

This enables performing more precise queries on general metrics, e.g.
from Grafana. One can quickly filter out all entries that are not
running, or that are templated, and use the name of the resource
as legend entry instead of the id, by imposing a filter on labels.